\name{bin}
\alias{bin}
\title{binary representation of x (an element from 1 to m) with m components}
\usage{
bin(x, m)
}
\arguments{
\item{x}{a positive integer less than or equal to m}

\item{m}{number of components, a positive integer}
}
\value{
a vector with elements 1 and 0
}
\description{
binary representation of x (an element from 1 to m) with m components
}
\examples{
bin(x = 2, m = 4);
}
