% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doofa.R
\name{coa}
\alias{coa}
\title{construct a component orthogonal array with m components}
\usage{
coa(m, lambda, ntrial)
}
\arguments{
\item{m}{a positive integer, currently supports less than 8}

\item{lambda}{a positive integer, usually 1}
\item{ntrial}{a positive integer, default is 10}
}
\value{
a component orthogonal array with m components
}
\description{
construct a component orthogonal array with m components such that each pair of columns contains each (i != j) combinations lambda times
}
\examples{
coa(m = 4, lambda = 1, ntrial = 10);
}
