% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doofa.R
\name{cycle}
\alias{cycle}
\title{cycle elements of a vector by one element to right}
\usage{
cycle(x)
}
\arguments{
\item{x}{a vector}
}
\value{
cycled vector
}
\description{
cycle elements of a vector by one element to right
}
\examples{
cycle(c(1, 2, 3));
}
