% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doofa.R
\name{doofa.pwo}
\alias{doofa.pwo}
\title{construct a design for order-of-addition experiment under pwo model with n runs and m components}
\usage{
doofa.pwo(n, m)
}
\arguments{
\item{n}{a  positive integer, preferably less than 100}

\item{m}{a positive integer, currently supports less than 8}
}
\value{
a design for order-of-addition experiment under pwo model with n runs and m components
}
\description{
construct a design for order-of-addition experiment under pwo model with n runs and m components
}
\examples{
doofa.pwo(5,3);
}
