% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doofa.R
\name{initial.design}
\alias{initial.design}
\title{create an initial design of o-of-a with n rows and m columns}
\usage{
initial.design(n, m)
}
\arguments{
\item{n}{a positive integer}

\item{m}{a positive integer}
}
\value{
a matrix with n rows and m columns
}
\description{
create an initial design of o-of-a with n rows and m columns
}
\examples{
initial.design(n = 6, m = 3);
}
