% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doofa.R
\name{one}
\alias{one}
\title{create a matrix of 1s with t rows}
\usage{
one(t)
}
\arguments{
\item{t}{a positive integer}
}
\value{
a matrix of 1s with t rows
}
\description{
create a matrix of 1s with t rows
}
\examples{
one(3);
}
