% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doofa.R
\name{pwo}
\alias{pwo}
\title{create PWO ordering of a given run of a design}
\usage{
pwo(x)
}
\arguments{
\item{x}{a numeric vector containing elements 1 to m in some order}
}
\value{
PWO ordering of the given run
}
\description{
create PWO ordering of the given run
}
\examples{
row = c(3,1,2)
pwo(row);
}
