% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doofa.R
\name{revbin}
\alias{revbin}
\title{reverse of bin function i.e., returns which elements of a binary vector is 1}
\usage{
revbin(x)
}
\arguments{
\item{x}{a vector with 0 and 1s such that there is only 1}
}
\value{
a postive integer m
}
\description{
reverse of bin function i.e., returns which elements of a binary vector is 1
}
\examples{
revbin(c(0,1,0,0));
}
