% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doofa.R
\name{shuffle}
\alias{shuffle}
\title{shuffle elements of a randomly chosen row of x matrix}
\usage{
shuffle(x)
}
\arguments{
\item{x}{a  matrix}
}
\value{
a matrix with shuffled elements of a row
}
\description{
shuffle elements of a randomly chosen row of x matrix
}
\examples{
x = matrix(c(3,1,2, 1,2,3,1,3,2,2,1,3),ncol = 3, byrow = TRUE)
shuffle(x);
}
