% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doofa.R
\name{swap}
\alias{swap}
\title{swap elements at i and i+1 of a vector}
\usage{
swap(x, i)
}
\arguments{
\item{x}{a vector}

\item{i}{a positive integer, less than length of x}
}
\value{
a vector with swapped elements
}
\description{
swap elements at i and i+1 of a vector
}
\examples{
swap(c(1,2,3),2);
}
