% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doofa.R
\name{vbin}
\alias{vbin}
\title{vectorized bin function}
\usage{
vbin(x)
}
\arguments{
\item{x}{a vector of length m with positive integers less than or equal to m}
}
\value{
a binary matrix
}
\description{
vectorized bin function
}
\examples{
vbin(c(3,1,2));
}
