% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doofa.R
\name{vrevbin}
\alias{vrevbin}
\title{vectorized revbin function}
\usage{
vrevbin(x, m)
}
\arguments{
\item{x}{a binary vector of length nm, such that each length of m has only one 1 and rest as 0}

\item{m}{a positive integer}
}
\value{
a vector of n positive integers
}
\description{
vectorized revbin function
}
\examples{
vrevbin(x=c(0,0,1,0,1,0), m = 3);
}
